bug fix to only delete button if it has post-id attribute

+++ public/deletePost.js
@@ -1,13 +1,15 @@
 const deletePost = async (event) => {
+  if (event.target.hasAttribute('post-id')){  
     event.preventDefault();
-    const postId = event.target.getAttribute('post-id');
-    const response = await fetch(`/api/post/${postId}`, {
-      method: 'DELETE'
-    });
-    if (response.ok) {
-      document.location.replace(`/dashboard`);
-    } else {
-      alert('Failed to delete post');
+      const postId = event.target.getAttribute('post-id');
+      const response = await fetch(`/api/post/${postId}`, {
+        method: 'DELETE'
+      });
+      if (response.ok) {
+        document.location.replace(`/dashboard`);
+      } else {
+        alert('Failed to delete post');
+      }
     }
   };
